+++ pathspec/__init__.py
@@ -44,7 +44,7 @@ __email__ = "cpburnz@gmail.com"
 __license__ = "MPL 2.0"
 __project__ = "pathspec"
 __status__ = "Development"
-__updated__ = "2017-10-05"
+__updated__ = "2018-03-30"
 __version__ = "0.5.6.dev1"
 
 from .pathspec import PathSpec